pyproject.toml
[project]
name = "ipml-boston-housing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cookiecutter-data-science>=2.3.0",
    "dvc[s3]>=3.64.2",
    "dvc-gdrive>=3.0.1",
    "pre-commit>=4.5.0",
    "ruff>=0.14.6",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "dvclive>=3.49.0",
    "loguru>=0.7.3",
    "scikit-learn>=1.7.2",
    "click>=8.3.1",
    "dvc-s3>=3.2.2",
    "mlflow[auth]>=3.7.0",
    "boto3>=1.41.5",
    "hydra-core>=1.3.2",
    "pydantic-settings>=2.0",
]

[dependency-groups]
dev = [
    "detect-secrets>=1.5.0",
]
